requirements.txt
fastapi
uvicorn[standard]
opencv-python
numpy
depthai
Pillow
python-multipart
pyserial
requests